+++ src/ParH2Matrix.h
@@ -719,7 +719,7 @@ sparse_matrix_type ParH2Matrix<Expansions,ColParticles>::gen_extended_matrix() c
         LOG(3,"\tnote: vector_size = "<<vector_size);
         LOG(3,"\tnote: size_W is = "<<size_W);
         LOG(3,"\tnote: size_g is = "<<size_g);
-        for (int i = 0; i < n; ++i) {
+        for (size_t i = 0; i < n; ++i) {
             LOG(4,"\tfor column "<<i<<", reserving "<<reserve[i]<<" rows");
         }
 
@@ -866,7 +866,7 @@ sparse_matrix_type ParH2Matrix<Expansions,ColParticles>::gen_extended_matrix() c
         A.makeCompressed();
 
 #ifndef NDEBUG
-        for (int i = 0; i < n; ++i) {
+        for (size_t i = 0; i < n; ++i) {
             size_t count = 0;
             for (sparse_matrix_type::InnerIterator it(A,i); it ;++it) {
                 count++;
@@ -1091,7 +1091,7 @@ sparse_matrix_type ParH2Matrix<Expansions,ColParticles>::gen_extended_matrix() c
         LOG(3,"\tnote: vector_size = "<<vector_size);
         LOG(3,"\tnote: size_W is = "<<size_W);
         LOG(3,"\tnote: size_g is = "<<size_g);
-        for (int i = 0; i < n; ++i) {
+        for (size_t i = 0; i < n; ++i) {
             LOG(4,"\tfor column "<<i<<", reserving "<<reserve[i]<<" rows");
         }
 
@@ -1148,7 +1148,7 @@ sparse_matrix_type ParH2Matrix<Expansions,ColParticles>::gen_extended_matrix() c
         A.makeCompressed();
 
 #ifndef NDEBUG
-        for (int i = 0; i < n; ++i) {
+        for (size_t i = 0; i < n; ++i) {
             size_t count = 0;
             for (sparse_matrix_type::InnerIterator it(A,i); it ;++it) {
                 count++;
